feat: tratamento de erros, e configurações na main.

+++ src/users/users.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repositories/users.repository';
+import { NotFoundError } from 'src/common/errors/types/NotFoundError';
 
 @Injectable()
 export class UsersService {
@@ -12,18 +13,36 @@ export class UsersService {
   }
 
   async findAll() {
-    return await this.repository.findAll();
+    const users = await this.repository.findAll();
+    if (users.length === 0) {
+      throw new NotFoundError('Nenhum usuário encontrado.');
+    }
+    return users;
   }
 
   async findOne(id: string) {
-    return await this.repository.findOne(id);
+    const user = await this.repository.findOne(id);
+    if (!user) {
+      throw new NotFoundError('Usuário não encontrado.');
+    }
+    return user;
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
+    const user = await this.repository.findOne(id);
+    if (!user) {
+      throw new NotFoundError('Usuário não encontrado.');
+    }
+
     return this.repository.update(id, updateUserDto);
   }
 
   async remove(id: string) {
+    const user = await this.repository.findOne(id);
+    if (!user) {
+      throw new NotFoundError('Usuário não encontrado.');
+    }
+
     return this.repository.remove(id);
   }
 }
